Implement levelOrder() method
We enter a while loop that continues as long as there are nodes in the queue.
Inside the loop, we dequeue a node from the front of the queue.
We push the value of the dequeued node to the result array.
If the dequeued node has a left child, we enqueue it to the queue.
If the dequeued node has a right child, we enqueue it to the queue.